feat(soulseek): add slskd HTTP adapter for routing through slskd REST API
When SLSKD_URL is set, Kima routes all Soulseek operations through
slskd's REST API instead of opening direct P2P connections. This lets
slskd handle network citizenship (file sharing, uploads, port
forwarding) while Kima drives search and download via HTTP.

The original P2P client is preserved and used when SLSKD_URL is unset,
so the change is fully backward-compatible.

- Rename client.ts -> p2p-client.ts (original vendored P2P client)
- Add slskd-client.ts (HTTP adapter targeting slskd REST API)
- New client.ts facade conditionally exports the right implementation

(cherry picked from commit f3cd59f2f2f17b1f9e372250f06ce2fcf8136c86)
